+++ src/card.tsx
@@ -59,6 +59,9 @@ export const useInnerText = (): {text: string; ref: (instance: HTMLElement | nul
 type BaseIconButtonAction = {
     Icon: (props: IconProps) => JSX.Element;
     label: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
+    'aria-current'?: React.AriaAttributes['aria-current'];
 };
 
 type IconButtonAction = BaseIconButtonAction &
@@ -202,20 +205,39 @@ type CardContainerProps = {
     dataAttributes?: DataAttributes;
     className?: string;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
 };
 
 const CardContainer = React.forwardRef<HTMLDivElement, CardContainerProps>(
     (
-        {children, width, height, aspectRatio, dataAttributes, className, 'aria-label': ariaLabel},
+        {
+            children,
+            width,
+            height,
+            aspectRatio,
+            dataAttributes,
+            className,
+            'aria-label': ariaLabel,
+            'aria-labelledby': ariaLabelledby,
+            'aria-description': ariaDescription,
+            'aria-describedby': ariaDescribedby,
+        },
         ref
     ): JSX.Element => {
         const cssAspectRatio = width && height ? undefined : aspectRatioToNumber(aspectRatio);
 
         return (
+            // aria-description should be vaild, but this eslint rule is complaining about it
+            // eslint-disable-next-line jsx-a11y/role-supports-aria-props
             <section
                 {...getPrefixedDataAttributes(dataAttributes)}
                 ref={ref}
                 aria-label={ariaLabel}
+                aria-labelledby={ariaLabelledby}
+                aria-description={ariaDescription}
+                aria-describedby={ariaDescribedby}
                 className={classNames(className, styles.cardContainer)}
                 style={{
                     width: width || '100%',
@@ -534,6 +556,7 @@ const CardContent = ({
 type TouchableProps = {
     trackingEvent?: TrackingEvent | ReadonlyArray<TrackingEvent>;
     role?: string;
+    'aria-current'?: React.AriaAttributes['aria-current'];
 } & ExclusifyUnion<
     | {
           href: string | undefined;
@@ -573,6 +596,9 @@ interface MediaCardBaseProps {
     children?: void;
     dataAttributes?: DataAttributes;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
     onClose?: () => void;
     closeButtonLabel?: string;
 }
@@ -608,6 +634,9 @@ export const MediaCard = React.forwardRef<HTMLDivElement, MediaCardProps>(
             buttonLink,
             dataAttributes,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             onClose,
             closeButtonLabel,
             ...touchableProps
@@ -620,18 +649,23 @@ export const MediaCard = React.forwardRef<HTMLDivElement, MediaCardProps>(
 
         const ariaLabel =
             ariaLabelProp ||
-            (isBiggerHeading(titleAs, pretitleAs)
-                ? [title, headlineText, pretitle, subtitle, description, extraText]
-                : [pretitle, headlineText, title, subtitle, description, extraText]
-            )
-                .filter(Boolean)
-                .join(' ');
+            (ariaLabeledByProp
+                ? undefined
+                : (isBiggerHeading(titleAs, pretitleAs)
+                      ? [title, headlineText, pretitle, subtitle, description, extraText]
+                      : [pretitle, headlineText, title, subtitle, description, extraText]
+                  )
+                      .filter(Boolean)
+                      .join(' '));
 
         return (
             <CardContainer
                 dataAttributes={{'component-name': 'MediaCard', testid: 'MediaCard', ...dataAttributes}}
                 ref={ref}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
                 className={styles.touchableContainer}
             >
                 <BaseTouchable
@@ -639,6 +673,9 @@ export const MediaCard = React.forwardRef<HTMLDivElement, MediaCardProps>(
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <Boxed className={styles.boxed} width="100%" height="100%">
                         {isTouchable && <div className={styles.touchableMediaCardOverlay} />}
@@ -723,6 +760,9 @@ export const NakedCard = React.forwardRef<HTMLDivElement, NakedCardProps>(
             buttonLink,
             dataAttributes,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             onClose,
             closeButtonLabel,
             ...touchableProps
@@ -736,18 +776,23 @@ export const NakedCard = React.forwardRef<HTMLDivElement, NakedCardProps>(
 
         const ariaLabel =
             ariaLabelProp ||
-            (isBiggerHeading(titleAs, pretitleAs)
-                ? [title, headlineText, pretitle, subtitle, description, extraText]
-                : [pretitle, headlineText, title, subtitle, description, extraText]
-            )
-                .filter(Boolean)
-                .join(' ');
+            (ariaLabeledByProp
+                ? undefined
+                : (isBiggerHeading(titleAs, pretitleAs)
+                      ? [title, headlineText, pretitle, subtitle, description, extraText]
+                      : [pretitle, headlineText, title, subtitle, description, extraText]
+                  )
+                      .filter(Boolean)
+                      .join(' '));
 
         return (
             <CardContainer
                 ref={ref}
                 dataAttributes={{'component-name': 'NakedCard', testid: 'NakedCard', ...dataAttributes}}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
                 className={isTouchable ? styles.touchableContainer : undefined}
             >
                 <BaseTouchable
@@ -755,6 +800,9 @@ export const NakedCard = React.forwardRef<HTMLDivElement, NakedCardProps>(
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <div className={styles.mediaCard} aria-hidden={isTouchable}>
                         <div style={{position: 'relative'}}>
@@ -823,6 +871,9 @@ type SmallNakedCardProps = MaybeTouchableCard<{
     extra?: React.ReactNode;
     dataAttributes?: DataAttributes;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
 }>;
 
 export const SmallNakedCard = React.forwardRef<HTMLDivElement, SmallNakedCardProps>(
@@ -839,6 +890,9 @@ export const SmallNakedCard = React.forwardRef<HTMLDivElement, SmallNakedCardPro
             extra,
             dataAttributes,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             ...touchableProps
         },
         ref
@@ -849,7 +903,10 @@ export const SmallNakedCard = React.forwardRef<HTMLDivElement, SmallNakedCardPro
         const {text: extraText, ref: extraRef} = useInnerText();
 
         const ariaLabel =
-            ariaLabelProp || [title, subtitle, description, extraText].filter(Boolean).join(' ');
+            ariaLabelProp ||
+            (ariaLabeledByProp
+                ? undefined
+                : [title, subtitle, description, extraText].filter(Boolean).join(' '));
 
         return (
             <CardContainer
@@ -860,6 +917,9 @@ export const SmallNakedCard = React.forwardRef<HTMLDivElement, SmallNakedCardPro
                     ...dataAttributes,
                 }}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
                 className={isTouchable ? styles.touchableContainer : undefined}
             >
                 <BaseTouchable
@@ -867,6 +927,9 @@ export const SmallNakedCard = React.forwardRef<HTMLDivElement, SmallNakedCardPro
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <div className={styles.mediaCard} aria-hidden={isTouchable}>
                         {media && (
@@ -957,6 +1020,9 @@ interface DataCardBaseProps {
     /** "data-" prefix is automatically added. For example, use "testid" instead of "data-testid" */
     dataAttributes?: DataAttributes;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
     onClose?: () => void;
     closeButtonLabel?: string;
 }
@@ -991,6 +1057,9 @@ export const DataCard = React.forwardRef<HTMLDivElement, DataCardProps>(
             buttonLink,
             dataAttributes,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             onClose,
             closeButtonLabel,
             aspectRatio,
@@ -1007,18 +1076,23 @@ export const DataCard = React.forwardRef<HTMLDivElement, DataCardProps>(
 
         const ariaLabel =
             ariaLabelProp ||
-            (isBiggerHeading(titleAs, pretitleAs)
-                ? [title, headlineText, pretitle, subtitle, description, extraText]
-                : [pretitle, headlineText, title, subtitle, description, extraText]
-            )
-                .filter(Boolean)
-                .join(' ');
+            (ariaLabeledByProp
+                ? undefined
+                : (isBiggerHeading(titleAs, pretitleAs)
+                      ? [title, headlineText, pretitle, subtitle, description, extraText]
+                      : [pretitle, headlineText, title, subtitle, description, extraText]
+                  )
+                      .filter(Boolean)
+                      .join(' '));
 
         return (
             <CardContainer
                 dataAttributes={{'component-name': 'DataCard', testid: 'DataCard', ...dataAttributes}}
                 ref={ref}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
                 className={styles.touchableContainer}
                 aspectRatio={aspectRatio}
             >
@@ -1027,6 +1101,9 @@ export const DataCard = React.forwardRef<HTMLDivElement, DataCardProps>(
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <Boxed className={styles.boxed} width="100%" height="100%">
                         {isTouchable && <div className={styles.touchableCardOverlay} />}
@@ -1107,6 +1184,9 @@ type SnapCardProps = MaybeTouchableCard<{
     /** "data-" prefix is automatically added. For example, use "testid" instead of "data-testid" */
     dataAttributes?: DataAttributes;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
     extra?: React.ReactNode;
     isInverse?: boolean;
     aspectRatio?: AspectRatio | number;
@@ -1126,6 +1206,9 @@ export const SnapCard = React.forwardRef<HTMLDivElement, SnapCardProps>(
             descriptionLinesMax,
             dataAttributes,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             extra,
             isInverse = false,
             aspectRatio,
@@ -1139,7 +1222,10 @@ export const SnapCard = React.forwardRef<HTMLDivElement, SnapCardProps>(
         const {text: extraText, ref: extraRef} = useInnerText();
 
         const ariaLabel =
-            ariaLabelProp || [title, subtitle, description, extraText].filter(Boolean).join(' ');
+            ariaLabelProp ||
+            (ariaDescribedByProp
+                ? undefined
+                : [title, subtitle, description, extraText].filter(Boolean).join(' '));
 
         return (
             <CardContainer
@@ -1148,12 +1234,18 @@ export const SnapCard = React.forwardRef<HTMLDivElement, SnapCardProps>(
                 className={styles.touchableContainer}
                 aspectRatio={aspectRatio}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
             >
                 <BaseTouchable
                     maybe
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <Boxed
                         className={styles.boxed}
@@ -1351,6 +1443,9 @@ interface CommonDisplayCardProps {
     description?: string;
     descriptionLinesMax?: number;
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
     aspectRatio?: AspectRatio | number;
     extra?: React.ReactNode;
 }
@@ -1426,6 +1521,9 @@ const DisplayCard = React.forwardRef<HTMLDivElement, GenericDisplayCardProps>(
             height,
             aspectRatio,
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             ...touchableProps
         },
         ref
@@ -1456,12 +1554,14 @@ const DisplayCard = React.forwardRef<HTMLDivElement, GenericDisplayCardProps>(
 
         const ariaLabel =
             ariaLabelProp ||
-            (isBiggerHeading(titleAs, pretitleAs)
-                ? [title, headlineText, pretitle, description, extraText]
-                : [pretitle, headlineText, title, description, extraText]
-            )
-                .filter(Boolean)
-                .join(' ');
+            (ariaLabeledByProp
+                ? undefined
+                : (isBiggerHeading(titleAs, pretitleAs)
+                      ? [title, headlineText, pretitle, description, extraText]
+                      : [pretitle, headlineText, title, description, extraText]
+                  )
+                      .filter(Boolean)
+                      .join(' '));
 
         return (
             <CardContainer
@@ -1471,6 +1571,9 @@ const DisplayCard = React.forwardRef<HTMLDivElement, GenericDisplayCardProps>(
                 height={height}
                 aspectRatio={aspectRatio}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
                 className={styles.touchableContainer}
             >
                 <BaseTouchable
@@ -1478,6 +1581,9 @@ const DisplayCard = React.forwardRef<HTMLDivElement, GenericDisplayCardProps>(
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <InternalBoxed
                         borderRadius={vars.borderRadii.legacyDisplay}
@@ -1644,6 +1750,9 @@ export const DisplayDataCard = React.forwardRef<HTMLDivElement, DisplayDataCardP
 
 interface PosterCardBaseProps {
     'aria-label'?: string;
+    'aria-labelledby'?: string;
+    'aria-description'?: string;
+    'aria-describedby'?: string;
     aspectRatio?: AspectRatio | number;
     width?: number | string;
     height?: number | string;
@@ -1703,6 +1812,9 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
             height,
             aspectRatio = '7:10',
             'aria-label': ariaLabelProp,
+            'aria-labelledby': ariaLabeledByProp,
+            'aria-description': ariaDescriptionProp,
+            'aria-describedby': ariaDescribedByProp,
             actions,
             onClose,
             closeButtonLabel,
@@ -1775,12 +1887,14 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
 
         const ariaLabel =
             ariaLabelProp ||
-            (isBiggerHeading(titleAs, pretitleAs)
-                ? [title, headlineText, pretitle, subtitle, description, extraText]
-                : [pretitle, headlineText, title, subtitle, description, extraText]
-            )
-                .filter(Boolean)
-                .join(' ');
+            (ariaLabeledByProp
+                ? undefined
+                : (isBiggerHeading(titleAs, pretitleAs)
+                      ? [title, headlineText, pretitle, subtitle, description, extraText]
+                      : [pretitle, headlineText, title, subtitle, description, extraText]
+                  )
+                      .filter(Boolean)
+                      .join(' '));
 
         return (
             <CardContainer
@@ -1791,12 +1905,18 @@ export const PosterCard = React.forwardRef<HTMLDivElement, PosterCardProps>(
                 aspectRatio={aspectRatio}
                 className={styles.touchableContainer}
                 aria-label={isTouchable ? undefined : ariaLabelProp}
+                aria-labelledby={isTouchable ? undefined : ariaLabeledByProp}
+                aria-description={isTouchable ? undefined : ariaDescriptionProp}
+                aria-describedby={isTouchable ? undefined : ariaDescribedByProp}
             >
                 <BaseTouchable
                     maybe
                     {...touchableProps}
                     className={styles.touchable}
                     aria-label={isTouchable ? ariaLabel : undefined}
+                    aria-labelledby={isTouchable ? ariaLabeledByProp : undefined}
+                    aria-description={isTouchable ? ariaDescriptionProp : undefined}
+                    aria-describedby={isTouchable ? ariaDescribedByProp : undefined}
                 >
                     <InternalBoxed
                         borderRadius={vars.borderRadii.legacyDisplay}